Store money as BigDecimal instead of Double
Floating point cannot represent most decimal amounts exactly (0.1 + 0.2 != 0.3),
so balances drift over many transfers. Balances and amounts are now BigDecimal
mapped to NUMERIC(19,2), compared with compareTo, and limited to 2 decimal places.

+++ src/main/java/com/example/payflow/controller/UserController.java
@@ -5,6 +5,7 @@ import com.example.payflow.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -38,7 +39,7 @@ public class UserController {
     }
 
     @GetMapping("/balance/{balance}")
-    public List<User> getUsersByBalance(@PathVariable Double balance) {
+    public List<User> getUsersByBalance(@PathVariable BigDecimal balance) {
         return userService.findByBalanceGreaterThanEqual(balance);
     }
 }
